+++ tests/js/spec/components/dropdownLink.spec.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import {shallow, mount} from 'enzyme';
+import TestUtils from 'react-addons-test-utils';
+import {shallow} from 'enzyme';
 import DropdownLink from 'app/components/dropdownLink';
 
 describe('DropdownLink', function() {
@@ -11,178 +12,39 @@ describe('DropdownLink', function() {
     menuClasses: ''
   };
 
-  describe('jQuery event listeners', function() {
-    let wrapper;
+  describe('componentWillUnmount()', function() {
+    it('should remove event handlers before unmounting', function() {
+      let dropdownlink = TestUtils.renderIntoDocument(<DropdownLink {...INPUT_1} />);
 
-    beforeEach(function() {
-      if (wrapper) {
-        wrapper.unmount();
-      }
-      jQuery(document).off('click');
+      let handlers = jQuery._data(dropdownlink.refs.dropdownToggle.parentNode, 'events');
+      expect(handlers).toBeInstanceOf(Object);
 
-      wrapper = mount(<DropdownLink title="test"><li>hi</li></DropdownLink>);
-    });
+      dropdownlink.componentWillUnmount(dropdownlink);
 
-    it('has document event listeners only when dropdown menu is open', function() {
-      let events = jQuery._data(document, 'events');
-      expect(events).toBeUndefined();
-
-      // Open
-      wrapper.find('a').simulate('click');
-      events = jQuery._data(document, 'events');
-      expect(events.click.length).toBe(1);
-
-      // Close
-      wrapper.find('a').simulate('click');
-      events = jQuery._data(document, 'events');
-      expect(events).toBeUndefined();
-    });
-
-    it('clears event listeners after unmount', function() {
-      let events = jQuery._data(document, 'events');
-      expect(events).toBeUndefined();
-
-      // Open
-      wrapper.find('a').simulate('click');
-      events = jQuery._data(document, 'events');
-      expect(events.click.length).toBe(1);
-
-      // Close
-      wrapper.unmount();
-      events = jQuery._data(document, 'events');
-      expect(events).toBeUndefined();
+      handlers = jQuery._data(dropdownlink.refs.dropdownToggle.parentNode, 'events');
+      expect(handlers).toBe(undefined);
     });
   });
 
-  describe('renders', function() {
-    it('and anchors to left by default', function() {
-      let component = shallow(
-        <DropdownLink {...INPUT_1}>
-          <div>1</div>
-          <div>2</div>
-        </DropdownLink>
-      );
+  it('renders and anchors to left by default', function() {
+    let component = shallow(
+      <DropdownLink {...INPUT_1}>
+        <div>1</div>
+        <div>2</div>
+      </DropdownLink>
+    );
 
-      expect(component).toMatchSnapshot();
-    });
-
-    it('and anchors to right', function() {
-      let component = shallow(
-        <DropdownLink {...INPUT_1} anchorRight>
-          <div>1</div>
-          <div>2</div>
-        </DropdownLink>
-      );
-
-      expect(component).toMatchSnapshot();
-    });
+    expect(component).toMatchSnapshot();
   });
 
-  describe('Uncontrolled', function() {
-    let wrapper;
+  it('renders and anchors to right', function() {
+    let component = shallow(
+      <DropdownLink {...INPUT_1} anchorRight>
+        <div>1</div>
+        <div>2</div>
+      </DropdownLink>
+    );
 
-    beforeEach(function() {
-      if (wrapper) {
-        wrapper.unmount();
-      }
-
-      wrapper = mount(<DropdownLink title="test"><li>hi</li></DropdownLink>);
-    });
-
-    describe('While Closed', function() {
-      it('displays dropdown menu when dropdown actor button clicked', function() {
-        expect(wrapper.find('li').length).toBe(0);
-        expect(wrapper.state('isOpen')).toBe(false);
-
-        // open
-        wrapper.find('a').simulate('click');
-        expect(wrapper.state('isOpen')).toBe(true);
-        expect(wrapper.find('li').length).toBe(1);
-      });
-    });
-    describe('While Opened', function() {
-      beforeEach(function() {
-        // Opens dropdown menu
-        wrapper.find('a').simulate('click');
-      });
-
-      it('closes when clicked outside', function() {
-        jQuery(document).click();
-        expect(wrapper.state('isOpen')).toBe(false);
-      });
-
-      it('closes when dropdown actor button is clicked', function() {
-        wrapper.find('a').simulate('click');
-        expect(wrapper.state('isOpen')).toBe(false);
-      });
-
-      it('closes when dropdown menu item is clicked', function() {
-        wrapper.find('li').simulate('click');
-        expect(wrapper.state('isOpen')).toBe(false);
-      });
-
-      it('does not close when menu is clicked and `keepMenuOpen` is on', function() {
-        wrapper = mount(
-          <DropdownLink title="test" keepMenuOpen><li>hi</li></DropdownLink>
-        );
-        wrapper.find('a').simulate('click');
-        wrapper.find('li').simulate('click');
-        expect(wrapper.state('isOpen')).toBe(true);
-        wrapper.unmount();
-      });
-    });
-  });
-
-  describe('Controlled', function() {
-    let wrapper;
-
-    beforeEach(function() {
-      if (wrapper) {
-        wrapper.unmount();
-      }
-    });
-    describe('Opened', function() {
-      beforeEach(function() {
-        wrapper = mount(
-          <DropdownLink isOpen={true} title="test"><li>hi</li></DropdownLink>
-        );
-      });
-
-      it('does not close when menu is clicked', function() {
-        // open
-        wrapper.find('li').simulate('click');
-        // State does not change
-        expect(wrapper.state('isOpen')).toBe(false);
-        expect(wrapper.find('.dropdown-menu').length).toBe(1);
-      });
-
-      it('does not close when document is clicked', function() {
-        jQuery(document).click();
-        // State does not change
-        expect(wrapper.state('isOpen')).toBe(false);
-        expect(wrapper.find('.dropdown-menu').length).toBe(1);
-      });
-
-      it('does not close when dropdown actor is clicked', function() {
-        wrapper.find('a').simulate('click');
-        // State does not change
-        expect(wrapper.state('isOpen')).toBe(false);
-        expect(wrapper.find('.dropdown-menu').length).toBe(1);
-      });
-    });
-    describe('Closed', function() {
-      beforeEach(function() {
-        wrapper = mount(
-          <DropdownLink isOpen={false} title="test"><li>hi</li></DropdownLink>
-        );
-      });
-
-      it('does not open when dropdown actor is clicked', function() {
-        wrapper.find('a').simulate('click');
-        // State does not change
-        expect(wrapper.state('isOpen')).toBe(false);
-        expect(wrapper.find('.dropdown-menu').length).toBe(0);
-      });
-    });
+    expect(component).toMatchSnapshot();
   });
 });
